+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-*None*
+- Constraint rollbacks are now being introspected based on the `pg_constraint` table
 
 ### Deprecated